requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
lupa==2.6
mwparserfromhell==0.7.2
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
